// src/generated/models/PlannerModel.ts
export interface GetTask_Response_V2 {
  createdBy?: {
    user?: {
      // The display name of the user who created this task.
      displayName?: string;
      // The id of the user who created this task.
      id?: string;
    };
  };
  // The id of the plan this task belongs to.
  planId?: string;
  // The id of the bucket this task belongs to.
  bucketId?: string;
  // The title of the task.
  title?: string;
  // The completion percentage of the task.
  percentComplete?: number;
  // The start datetime of the task.
  startDateTime?: string;
  // The datetime the task was created.
  createdDateTime?: string;
  // The datetime the task is due.
  dueDateTime?: string;
  // Set to true if the task has a description.
  hasDescription?: boolean;
  // The datetime the task was completed.
  completedDateTime?: string;
  // The number of external references that exist on the task.
  referenceCount?: number;
  // The id of the task.
  id?: string;
  appliedCategories?: AppliedCategories;
  _assignments?: {
   // The id of the user to whom this task was assigned to.
   userId?: string;
   value?: {
     assignedBy?: {
       user?: {
         // The id of the user who assigned this task.
         id?: string;
       };
     };
     // The datetime the task was assigned.
     assignedDateTime?: string;
     // Order Hint
     orderHint?: string;
   };
 }[];
}
// ...
export interface AppliedCategories {
  // True if the task has the Pink category.
  category1?: boolean;
  // True if the task has the Red category.
  category2?: boolean;
  // True if the task has the Yellow category.
  category3?: boolean;
  // True if the task has the Green category.
  category4?: boolean;
  // True if the task has the Blue category.
  category5?: boolean;
  // True if the task has the Purple category.
  category6?: boolean;
  // True if the task has the Bronze category.
  category7?: boolean;
  // True if the task has the Lime category.
  category8?: boolean;
  // True if the task has the Aqua category.
  category9?: boolean;
  // True if the task has the Gray category.
  category10?: boolean;
  // True if the task has the Silver category.
  category11?: boolean;
  // True if the task has the Brown category.
  category12?: boolean;
  // True if the task has the Cranberry category.
  category13?: boolean;
  // True if the task has the Orange category.
  category14?: boolean;
  // True if the task has the Peach category.
  category15?: boolean;
  // True if the task has the Marigold category.
  category16?: boolean;
  // True if the task has the Light green category.
  category17?: boolean;
  // True if the task has the Dark green category.
  category18?: boolean;
  // True if the task has the Teal category.
  category19?: boolean;
  // True if the task has the Light blue category.
  category20?: boolean;
  // True if the task has the Dark blue category.
  category21?: boolean;
  // True if the task has the Lavender category.
  category22?: boolean;
  // True if the task has the Plum category.
  category23?: boolean;
  // True if the task has the Light gray category.
  category24?: boolean;
  // True if the task has the Dark gray category.
  category25?: boolean;
}